Moved streams functions in stream-array
